fixed several warnings generated in ProjectBuilder [Michael D. Adams]
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1694 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Expat/xmlparse.c
@@ -941,7 +941,7 @@ const XML_LChar *XML_ErrorString(int code)
     XML_T("error in processing external entity reference"),
     XML_T("document is not standalone")
   };
-  if (code > 0 && code < sizeof(message)/sizeof(message[0]))
+  if (code > 0 && (size_t)(code) < sizeof(message)/sizeof(message[0]))
     return message[code];
   return 0;
 }
